Phase 8.3.8: Fix no-any-return type errors (-19 errors)
Added float() and np.asarray() casts to handle numpy operations that return Any.
Added cast() for DataFrame returns from pandas operations.

Changes:
- statistical.py: Added float() casts for scipy stats functions
- matrix_ops.py: Added np.asarray() for vcov calculation
- standard_errors/utils.py: Added np.asarray() for 4 functions
- newey_west.py, driscoll_kraay.py: Added float() casts for kernel weights
- gmm/estimator.py: Added np.asarray() and bool() casts (4 functions)
- core/panel_data.py: Added cast(pd.DataFrame, ...) for 4 methods

Progress: 205 → 186 errors (-19 errors, 9.3% reduction)
Cumulative: 233 → 186 errors (-47 errors, 20.2% total reduction)

+++ panelbox/utils/statistical.py
@@ -49,9 +49,9 @@ def compute_pvalue(tstat: float, df: int, two_sided: bool = True) -> float:
         p-value
     """
     if two_sided:
-        return 2 * (1 - stats.t.cdf(np.abs(tstat), df))
+        return float(2 * (1 - stats.t.cdf(np.abs(tstat), df)))
     else:
-        return 1 - stats.t.cdf(tstat, df)
+        return float(1 - stats.t.cdf(tstat, df))
 
 
 def compute_fstat(
@@ -161,4 +161,4 @@ def compute_chi2_pvalue(statistic: float, df: int) -> float:
     float
         p-value
     """
-    return 1 - stats.chi2.cdf(statistic, df)
+    return float(1 - stats.chi2.cdf(statistic, df))
